+++ src/bioimageio/core/utils/_type_guards.py
@@ -6,3 +6,5 @@ from bioimageio.spec._internal import type_guards
 is_list = type_guards.is_list
 is_ndarray = type_guards.is_ndarray
 is_tuple = type_guards.is_tuple
+is_dict = type_guards.is_dict
+is_kwargs = type_guards.is_kwargs